Engine now can execute functions with args
This commit is the most raw commit in this project. Viewer discretion is
adviced. It was written so fast that some classes even doesn't have any
destructors! DONT TRY TO READ IT. Look some older versions that are ok,
or newer commits which fixed this bullshit code. MATURE CONTENT INSIDE
THIS COMMIT.