3rd iteration of refactoring
run.h file is now main file that starts functions from compile.h and
execute.h files, which is more logical. All files contain functions not
bigger than 80 lines, which is good, but not best. Code is divided in
function good. All functions are logically good named and calling each
other in obvious way. but arguments still need to be worked all.

and reliability is poor.